Akey, this is required to create the Library constructor (for reference)
https://stackoverflow.com/questions/17240708/implicit-super-constructor-object-is-undefined-must-explicitly-invoke-another/19570719

What you have in this push is the skeleton of the programme, the rest is
just about adding flesh (functions). It's basically the Library box from
the Library_Project_Master (page 4) diagram. I will double check it with
my friend tomorrow but I think it's a good start. I think it will be
important to follow the development order starting from Member (add and
get).

I don't know what aggregation and inheritance signs mean in practice on
the diagram

https://www.google.com/url?sa=i&url=https%3A%2F%2Fen.wikipedia.org%2Fwiki%2FClass_diagram&psig=AOvVaw1phKjDM0mIXBV6NGZwJ97M&ust=1606519748534000&source=images&cd=vfe&ved=0CAIQjRxqFwoTCPizx8-uoe0CFQAAAAAdAAAAABAD